Trait copypasta::ClipboardProvider

pub trait ClipboardProvider: Send {
    fn get_contents(&mut self) -> Result<String, Box<dyn Error>>;

    fn set_contents(&mut self, _: String) -> Result<(), Box<dyn Error>>;
}

Trait for clipboard access

Required methods

fn get_contents(&mut self) -> Result<String, Box<dyn Error>>

Method to get the clipboard contents as a String

fn set_contents(&mut self, _: String) -> Result<(), Box<dyn Error>>

Method to set the clipboard contents as a String

Loading content...