Function copypasta::wayland_clipboard::create_clipboards_from_external

pub unsafe fn create_clipboards_from_external(
    display: *mut c_void
) -> (Primary, Clipboard)

Create new clipboard from a raw display pointer.

Safety

Since the type of the display is a raw pointer, it's the responsibility of the callee to make sure that the passed pointer is a valid Wayland display.